+++ src/collect/adding/addAcceptedIssues.ts
@@ -1,3 +1,5 @@
+import { descriptionToCoAuthors } from "description-to-co-authors";
+
 import { ContributorsCollection } from "../../ContributorsCollection.js";
 import { AllContributorsForRepositoryOptions } from "../../options.js";
 import { AcceptedIssue } from "../collecting/collectAcceptedIssues.js";
@@ -25,12 +27,24 @@ export function addAcceptedIssues(
 			// - 🔧 `tool`: authors of merged PRs that address issues labeled as tooling
 			[options.labelTypeTool, "tool"],
 		] as const) {
-			if (labels.some((label) => label === labelType)) {
-				contributors.add(
-					acceptedIssue.user?.login,
-					acceptedIssue.number,
-					contribution,
-				);
+			if (!labels.some((label) => label === labelType)) {
+				continue;
+			}
+
+			const logins = [];
+
+			if (acceptedIssue.user) {
+				logins.push(acceptedIssue.user.login);
+			}
+
+			if (acceptedIssue.body) {
+				for (const coAuthor of descriptionToCoAuthors(acceptedIssue.body)) {
+					logins.push(coAuthor.username);
+				}
+			}
+
+			for (const login of logins) {
+				contributors.add(login, acceptedIssue.number, contribution);
 			}
 		}
 	}