+++ src/index.ts
@@ -3,3 +3,4 @@ export {
 	type ContributorsContributions,
 } from "./ContributorsCollection.js";
 export { getAllContributorsForRepository } from "./getAllContributorsForRepository.js";
+export type { RawAllContributorsForRepositoryOptions } from "./options.js";