+++ src/options.ts
@@ -6,6 +6,8 @@ const defaultOptions = {
 		"copilot[bot]",
 		"dependabot",
 		"dependabot[bot]",
+		"github-actions",
+		"github-actions[bot]",
 		"renovate",
 		"renovate[bot]",
 	],
@@ -15,6 +17,18 @@ const defaultOptions = {
 	labelTypeIdeas: "type: feature",
 	labelTypeTool: "area: tooling",
 };
+
+export interface AllContributorsForRepositoryOptions {
+	auth?: string;
+	ignoredLogins: Set<string>;
+	labelAcceptingPrs: string;
+	labelTypeBug: string;
+	labelTypeDocs: string;
+	labelTypeIdeas: string;
+	labelTypeTool: string;
+	owner: string;
+	repo: string;
+}
 
 export interface RawAllContributorsForRepositoryOptions {
 	/**
@@ -60,18 +74,6 @@ export interface RawAllContributorsForRepositoryOptions {
 	/**
 	 *  The name of the repository to query, such as `"all-contributors-for-repository"`.
 	 */
-	repo: string;
-}
-
-export interface AllContributorsForRepositoryOptions {
-	auth?: string;
-	ignoredLogins: Set<string>;
-	labelAcceptingPrs: string;
-	labelTypeBug: string;
-	labelTypeDocs: string;
-	labelTypeIdeas: string;
-	labelTypeTool: string;
-	owner: string;
 	repo: string;
 }